Fixes #5
The library uses special functions for symmetric, time-invariant boundaries.
This specialization happens in DMBase::create and first makes sure that
the boundaries are constant (i.e., that the b_lo and b_up vectors have
length 1) and that the first element of b_lo is the same as that of b_up,
except for a sign change.

A bug in the code didn't compare b_lo to b_up, but instead b_up to b_up.
This commit fixes this issue.